Move persistIdentity to after the wrapped middleware runs
Persisting the identity before calling $next results in the session
being rotated before SecurityComponent handles the request. By doing the
rotation after we avoid that problem.

Moving when the session is updated will affect applications as the
session data will *not* be populated during the controller action. This
could be a breaking change but I think it is required to keep the
session rotation fixes in place.