Fix logout with session authentication
When using the session authenticator one could never logout. Because the
session authenticator would be the 'authentication provider' for the
current request, it would have data re-persisted by the middleware after
the wrapped middleware (including the controller) were complete. By
clearing the authentication provider we skip this loop.

Backport #265 to master